Add create, update and delete opportunity for task's instances
