+++ manager/admin.py
