+++ manager/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class TaskConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "task"
+    name = "manager"
