+++ manager/urls.py
@@ -1,0 +1,12 @@
+from django.urls import path
+
+from .views import TaskListView, TaskCreateView, TaskUpdateView, TaskDeleteView
+
+urlpatterns = [
+    path("", TaskListView.as_view(), name="task-list"),
+    path("tasks/create/", TaskCreateView.as_view(), name="task-create"),
+    path("tasks/<int:pk>/update/", TaskUpdateView.as_view(), name="task-update"),
+    path("tasks/<int:pk>/delete/", TaskDeleteView.as_view(), name="task-delete"),
+]
+
+app_name = "manager"
