--- task/urls.py
@@ -1,9 +1,0 @@
-from django.urls import path
-
-from .views import TaskListView
-
-urlpatterns = [
-    path("", TaskListView.as_view(), name="index"),
-]
-
-app_name = "task"
