+++ templates/includes/sidebar.html
@@ -1,0 +1,15 @@
+<ul class="sidebar-nav list-group">
+  {% if user.is_authenticated %}
+    <li class="list-group-item">User: <a href="{{ user.get_absolute_url }}">{{ user.get_username }}</a></li>
+    <li class="list-group-item"><a href="#">Logout</a></li>
+  {% else %}
+    <li class="list-group-item"><a href="#">Login</a></li>
+  {% endif %}
+
+  <br>
+
+  <li class="list-group-item"><a href="#">Home</a></li>
+  <li class="list-group-item"><a href="#">All drivers</a></li>
+  <li class="list-group-item"><a href="#">All cars</a></li>
+  <li class="list-group-item"><a href="#">All Tasks</a></li>
+</ul>
