+++ templates/layouts/base.html
@@ -10,8 +10,10 @@
 <body>
   <div class="container-fluid">
     <div class="row">
-      <div class="col-sm-10 ">
-
+      <div class="col-sm-2 ">
+        {% block sidebar %}
+          {% include "includes/sidebar.html" %}
+        {% endblock %}
         {% block content %}{% endblock %}
 
       </div>
