+++ todoList/urls.py
@@ -4,5 +4,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("", include("task.urls", namespace="task")),
+    path("", include("manager.urls", namespace="manager")),
 ]
